Move the helper class to the integration package
